refactor: 모달 창 다른 컴포넌트로 분리
