+++ src/components/projects/settings/ManageRight.vue
@@ -51,56 +51,18 @@
       </div>
     </div>
 
-    <!-- 멤버 추가 모달 -->
-    <div v-if="showAddMemberModal" class="modal-overlay" @click="closeModal">
-      <div class="modal" @click.stop>
-        <div class="modal-header">
-          <h3 class="modal-title">새 멤버 추가</h3>
-          <button @click="closeModal" class="modal-close">✕</button>
-        </div>
-
-        <div class="modal-body">
-          <div class="form-group">
-            <label class="form-label">이름</label>
-            <input
-              v-model="newMember.name"
-              type="text"
-              class="form-input"
-              placeholder="멤버 이름을 입력하세요"
-            />
-          </div>
-
-          <div class="form-group">
-            <label class="form-label">부서</label>
-            <input
-              v-model="newMember.department"
-              type="text"
-              class="form-input"
-              placeholder="부서명을 입력하세요"
-            />
-          </div>
-
-          <div class="form-group">
-            <label class="form-label">권한</label>
-            <select v-model="newMember.permission" class="form-select">
-              <option value="Read">Read</option>
-              <option value="Read/Write">Read/Write</option>
-              <option value="Admin">Admin</option>
-            </select>
-          </div>
-        </div>
-
-        <div class="modal-footer">
-          <button @click="closeModal" class="btn-cancel">취소</button>
-          <button @click="addMember" class="btn-add">추가</button>
-        </div>
-      </div>
-    </div>
+    <!-- 멤버 추가 모달 컴포넌트 -->
+    <AddMemberModal
+      :is-visible="showAddMemberModal"
+      @close="showAddMemberModal = false"
+      @add-member="handleAddMember"
+    />
   </div>
 </template>
 
 <script setup>
-import { ref, reactive } from "vue";
+import { ref } from "vue";
+import AddMemberModal from "./AddMemberModal.vue";
 
 const showAddMemberModal = ref(false);
 
@@ -131,12 +93,6 @@ const members = ref([
   },
 ]);
 
-const newMember = reactive({
-  name: "",
-  department: "",
-  permission: "Read",
-});
-
 const updatePermission = (memberId, permission) => {
   console.log(`멤버 ${memberId}의 권한을 ${permission}으로 변경`);
   // 실제 권한 업데이트 로직
@@ -148,30 +104,10 @@ const deleteMember = (index) => {
   }
 };
 
-const addMember = () => {
-  if (newMember.name && newMember.department) {
-    members.value.push({
-      id: Date.now(),
-      name: newMember.name,
-      department: newMember.department,
-      permission: newMember.permission,
-    });
-
-    // 폼 리셋
-    newMember.name = "";
-    newMember.department = "";
-    newMember.permission = "Read";
-
-    closeModal();
-  }
-};
-
-const closeModal = () => {
-  showAddMemberModal.value = false;
-  // 폼 리셋
-  newMember.name = "";
-  newMember.department = "";
-  newMember.permission = "Read";
+const handleAddMember = (newMember) => {
+  members.value.push(newMember);
+  console.log("새 멤버 추가:", newMember);
+  // 실제 멤버 추가 API 호출 로직
 };
 </script>
 
@@ -353,151 +289,7 @@ const closeModal = () => {
   font-size: 1rem;
 }
 
-/* 모달 스타일 */
-.modal-overlay {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 50;
-  backdrop-filter: blur(4px);
-  animation: fadeIn 0.3s ease-out;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-.modal {
-  background: white;
-  border-radius: 16px;
-  width: 90%;
-  max-width: 480px;
-  box-shadow: 0 20px 25px rgba(0, 0, 0, 0.1);
-  animation: modalSlideIn 0.3s ease-out;
-}
-
-@keyframes modalSlideIn {
-  from {
-    opacity: 0;
-    transform: translateY(-20px) scale(0.95);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0) scale(1);
-  }
-}
-
-.modal-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1.5rem;
-  border-bottom: 1px solid #f3f4f6;
-}
-
-.modal-title {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #111827;
-  margin: 0;
-}
-
-.modal-close {
-  background: transparent;
-  border: none;
-  color: #6b7280;
-  font-size: 1.25rem;
-  cursor: pointer;
-  padding: 0.25rem;
-  border-radius: 0.25rem;
-  transition: all 0.2s;
-}
-
-.modal-close:hover {
-  background: #f3f4f6;
-  color: #374151;
-}
-
-.modal-body {
-  padding: 1.5rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.form-group {
-  display: flex;
-  flex-direction: column;
-}
-
-.form-label {
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: #374151;
-  margin-bottom: 0.5rem;
-}
-
-.form-input,
-.form-select {
-  padding: 0.75rem 1rem;
-  border: 2px solid #e5e7eb;
-  border-radius: 8px;
-  font-size: 0.875rem;
-  transition: all 0.3s ease;
-}
-
-.form-input:focus,
-.form-select:focus {
-  outline: none;
-  border-color: #10b981;
-  box-shadow: 0 0 0 3px rgba(16, 185, 129, 0.1);
-}
-
-.modal-footer {
-  display: flex;
-  gap: 0.75rem;
-  justify-content: flex-end;
-  padding: 1.5rem;
-  border-top: 1px solid #f3f4f6;
-}
-
-.btn-cancel,
-.btn-add {
-  padding: 0.75rem 1.5rem;
-  border-radius: 8px;
-  font-size: 0.875rem;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.btn-cancel {
-  background: #f9fafb;
-  color: #374151;
-  border: 1px solid #d1d5db;
-}
-
-.btn-cancel:hover {
-  background: #f3f4f6;
-}
-
-.btn-add {
-  background: #10b981;
-  color: white;
-  border: 1px solid #10b981;
-}
-
-.btn-add:hover {
-  background: #059669;
-}
+/* 모달 관련 스타일은 AddMemberModal.vue로 이동 */
 
 /* 반응형 디자인 */
 @media (max-width: 768px) {
